Refactor task fetching in TaskDashboard and introduce API service
- Integrated a new TaskApiService for fetching user tasks with improved error handling and authentication.
- Removed direct API calls from the TaskDashboard component, enhancing code organization.
- Updated loading and error states for better user experience.
- Added UserTaskAssignment interface to types for better type safety.
